+++ api/writeAPIBlocking_test.go
@@ -6,56 +6,56 @@ package api
 
 import (
 	"context"
-
 	"sync"
 	"testing"
 	"time"
 
+	http2 "github.com/influxdata/influxdb-client-go/api/http"
 	"github.com/influxdata/influxdb-client-go/api/write"
-	"github.com/influxdata/influxdb-client-go/internal/http"
+	"github.com/influxdata/influxdb-client-go/internal/test"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
 func TestWritePoint(t *testing.T) {
-	service := newTestService(t, "http://localhost:8888")
+	service := test.NewTestService(t, "http://localhost:8888")
 	writeAPI := NewWriteAPIBlocking("my-org", "my-bucket", service, write.DefaultOptions().SetBatchSize(5))
 	points := genPoints(10)
 	err := writeAPI.WritePoint(context.Background(), points...)
 	require.Nil(t, err)
-	require.Len(t, service.lines, 10)
+	require.Len(t, service.Lines(), 10)
 	for i, p := range points {
 		line := write.PointToLineProtocol(p, writeAPI.writeOptions.Precision())
 		//cut off last \n char
 		line = line[:len(line)-1]
-		assert.Equal(t, service.lines[i], line)
+		assert.Equal(t, service.Lines()[i], line)
 	}
 }
 
 func TestWriteRecord(t *testing.T) {
-	service := newTestService(t, "http://localhost:8888")
+	service := test.NewTestService(t, "http://localhost:8888")
 	writeAPI := NewWriteAPIBlocking("my-org", "my-bucket", service, write.DefaultOptions().SetBatchSize(5))
 	lines := genRecords(10)
 	err := writeAPI.WriteRecord(context.Background(), lines...)
 	require.Nil(t, err)
-	require.Len(t, service.lines, 10)
+	require.Len(t, service.Lines(), 10)
 	for i, l := range lines {
-		assert.Equal(t, l, service.lines[i])
+		assert.Equal(t, l, service.Lines()[i])
 	}
 	service.Close()
 
 	err = writeAPI.WriteRecord(context.Background())
 	require.Nil(t, err)
-	require.Len(t, service.lines, 0)
+	require.Len(t, service.Lines(), 0)
 
-	service.replyError = &http.Error{Code: "invalid", Message: "data"}
+	service.SetReplyError(&http2.Error{Code: "invalid", Message: "data"})
 	err = writeAPI.WriteRecord(context.Background(), lines...)
 	require.NotNil(t, err)
 	require.Equal(t, "invalid: data", err.Error())
 }
 
 func TestWriteContextCancel(t *testing.T) {
-	service := newTestService(t, "http://localhost:8888")
+	service := test.NewTestService(t, "http://localhost:8888")
 	writeAPI := NewWriteAPIBlocking("my-org", "my-bucket", service, write.DefaultOptions().SetBatchSize(5))
 	lines := genRecords(10)
 	ctx, cancel := context.WithCancel(context.Background())
@@ -70,5 +70,5 @@ func TestWriteContextCancel(t *testing.T) {
 	cancel()
 	wg.Wait()
 	require.Equal(t, context.Canceled, err)
-	assert.Len(t, service.lines, 0)
+	assert.Len(t, service.Lines(), 0)
 }