+++ client_test.go
@@ -7,6 +7,7 @@ package influxdb2
 import (
 	"context"
 	"fmt"
+	http3 "github.com/influxdata/influxdb-client-go/api/http"
 	"net/http"
 	"net/http/httptest"
 	"testing"
@@ -45,6 +46,37 @@ func TestUrls(t *testing.T) {
 	}
 }
 
+func TestWriteAPIManagement(t *testing.T) {
+	data := []struct {
+		org          string
+		bucket       string
+		expectedCout int
+	}{
+		{"o1", "b1", 1},
+		{"o1", "b2", 2},
+		{"o1", "b1", 2},
+		{"o2", "b1", 3},
+		{"o2", "b2", 4},
+		{"o1", "b2", 4},
+		{"o1", "b3", 5},
+		{"o2", "b2", 5},
+	}
+	c := NewClient("http://localhost", "x").(*clientImpl)
+	for i, d := range data {
+		t.Run(fmt.Sprintf("%d", i), func(t *testing.T) {
+			w := c.WriteAPI(d.org, d.bucket)
+			assert.NotNil(t, w)
+			assert.Len(t, c.writeAPIs, d.expectedCout)
+			wb := c.WriteAPIBlocking(d.org, d.bucket)
+			assert.NotNil(t, wb)
+			assert.Len(t, c.syncWriteAPIs, d.expectedCout)
+		})
+	}
+	c.Close()
+	assert.Len(t, c.writeAPIs, 0)
+	assert.Len(t, c.syncWriteAPIs, 0)
+}
+
 func TestUserAgent(t *testing.T) {
 	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		time.Sleep(100 * time.Millisecond)
@@ -78,7 +110,7 @@ func TestServerError429(t *testing.T) {
 	c := NewClient(server.URL, "x")
 	err := c.WriteAPIBlocking("o", "b").WriteRecord(context.Background(), "a,a=a a=1i")
 	require.NotNil(t, err)
-	perror, ok := err.(*http2.Error)
+	perror, ok := err.(*http3.Error)
 	require.True(t, ok)
 	require.NotNil(t, perror)
 	assert.Equal(t, "too many requests", perror.Code)
@@ -112,7 +144,7 @@ func TestServerErrorNonJSON(t *testing.T) {
 	c := NewClient(server.URL, "x")
 	err := c.WriteAPIBlocking("o", "b").WriteRecord(context.Background(), "a,a=a a=1i")
 	require.NotNil(t, err)
-	perror, ok := err.(*http2.Error)
+	perror, ok := err.(*http3.Error)
 	require.True(t, ok)
 	require.NotNil(t, perror)
 	assert.Equal(t, "500 Internal Server Error", perror.Code)
@@ -131,7 +163,7 @@ func TestServerErrorInflux1_8(t *testing.T) {
 	c := NewClient(server.URL, "x")
 	err := c.WriteAPIBlocking("o", "b").WriteRecord(context.Background(), "a,a=a a=1i")
 	require.NotNil(t, err)
-	perror, ok := err.(*http2.Error)
+	perror, ok := err.(*http3.Error)
 	require.True(t, ok)
 	require.NotNil(t, perror)
 	assert.Equal(t, "404 Not Found", perror.Code)